Remove unused BaseRenderer base class.
